fixed saving fortunes_file in the quit menu option

+++ src/cassandra.rb
@@ -1,6 +1,4 @@
 #!/usr/bin/ruby
-
-# Main programme loop
 require_relative 'fortune'
 
 def pause()
@@ -16,6 +14,7 @@ fortunes_file = "./fortunes-from-cassandra.txt"
 
 puts "Hello and welcome"
 
+# Main programme loop
 loop do
   puts "Options:"
   puts "1) What's in my nearest future?"
@@ -30,7 +29,7 @@ loop do
       todays_fortune = Fortune.new
       pause()
       puts todays_fortune.quote
-      File.open(fortunes_file, "a") {|file| file.write("#{Time.now.strftime("%d of %B, %Y")} - #{todays_fortune.to_s}\n") }
+      File.open(fortunes_file, "a") {|file| file.write("#{Time.now.strftime("%d %B, %Y %H:%M")} - #{todays_fortune.to_s}\n") }
       puts "Would you like another? 1 yes 2 no"
       input = gets.strip.to_i
       break if input == 2
@@ -42,14 +41,25 @@ loop do
     if File.exist?(fortunes_file)
       pause()
       File.readlines(fortunes_file).map {|fortune| puts fortune}
+      puts "Save your fortunes? 1 yes 2 no"
+      input = gets.strip.to_i
+      if input == 2
+        File.delete("fortunes-from-cassandra.txt")
+      elsif input == 1
+        puts "Saved here: #{File.expand_path(File.dirname(fortunes_file))}"
+      end
     else
       puts "No fortunes yet. Let\'s give you your first one!"
     end
   when 3
-    puts "Save your fortunes? 1 yes 2 no"
-    input = gets.strip.to_i
-    if input == 2
-      File.delete("fortunes-from-cassandra.txt")
+    if File.exist?(fortunes_file)
+      puts "Save your fortunes? 1 yes 2 no"
+      input = gets.strip.to_i
+      if input == 2
+        File.delete("fortunes-from-cassandra.txt")
+      elsif input == 1
+        puts "Saved here: #{File.expand_path(File.dirname(fortunes_file))}"
+      end
     end
     exit
   else
